[plugin.video.odeon] 1.3.2
