+++ plugin.video.odeon/README.md
@@ -1,7 +1,7 @@
 ![CINE.AR](https://www.odeon.com.ar/favicon-96x96.png)
 
 # odeon-kodi
-Addon de CINE.AR PLAY / ODEON (www.odeon.com.ar) para Kodi.
+Addon de CINE.AR PLAY / ODEON para Kodi.
 
 ### Description
 
@@ -14,8 +14,6 @@ CINE.AR PLAY is the free video on demand platform where you will find the best A
 CINE.AR PLAY es la plataforma de video a demanda gratuita para ver todos los contenidos nacionales donde quieras, cuando quieras. Las películas y las series, los documentales y los cortos, las nuevas miradas y los clásicos de siempre para disfrutarlos desde cualquier dispositivo.
 
 ### Addtional information
-- http://www.cine.ar/play
-- https://www.odeon.com.ar
+- https://www.cine.ar/play
 - https://arsat.github.io/odeon-kodi/
 - http://forum.kodi.tv/showthread.php?tid=296742
-- Contact: dev@odeon.com.ar
