+++ plugin.video.odeon/changelog.txt
@@ -1,6 +1,10 @@
 ﻿# Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.3.2] - 2017-12-01
+### Changed
+- My Rentals should not be shown.
+
 ## [1.3.1] - 2017-07-07
 ### Changed
 - Minor bug fix.
